Create The_rook's_move
Two different squares of the chessboard are given. Write a program that determines whether a rook can get from the first square to the second one in one move. The program receives four numbers from 1 to 8 each, specifying the column number and row number first for the first cell, then for the second cell. The program should output "YES" if it is possible to get from the first cell to the second by the move of the rook, or "NO" otherwise.