Create The_King's_Move
Two different squares of the chessboard are given. Write a program that determines whether the king can get from the first square to the second one in one move. The program receives four numbers from 1 to 8 each, specifying the column number and row number first for the first cell, then for the second cell. The program should output "YES" if it is possible to get from the first square with the king's move to the second, or "NO" otherwise.